Refactor campaign execution worker into separate service classes
