+++ server/src/workers/campaign-execution/timeout-execution.service.ts
@@ -1,0 +1,132 @@
+import type { Job } from 'bullmq';
+import { logger } from '@/libs/logger';
+import { contactCampaignRepository, messageEventRepository } from '@/repositories';
+import { campaignPlanExecutionService } from '@/modules/campaign/campaignPlanExecution.service';
+import type { TimeoutJobPayload } from '@/types/timeout.types';
+
+export interface TimeoutJobResult {
+  success: boolean;
+  skipped?: boolean;
+  reason?: string;
+  syntheticEventId?: string;
+}
+
+export class TimeoutExecutionService {
+  async processTimeout(job: Job<TimeoutJobPayload>): Promise<TimeoutJobResult> {
+    const { campaignId, nodeId, messageId, eventType, tenantId } = job.data;
+
+    logger.info('[CampaignExecutionWorker] Processing timeout job', {
+      jobId: job.id,
+      campaignId,
+      nodeId,
+      eventType,
+      messageId,
+      tenantId,
+    });
+
+    try {
+      // Check if real event already happened
+      const realEventType = eventType.replace('no_', ''); // no_open -> open
+      const realEventExists = await messageEventRepository.findByMessageAndType(
+        messageId,
+        realEventType
+      );
+
+      if (realEventExists) {
+        logger.info('[CampaignExecutionWorker] Real event found, skipping synthetic event', {
+          messageId,
+          eventType,
+          realEventId: realEventExists.id,
+          jobId: job.id,
+        });
+        return {
+          success: true,
+          skipped: true,
+          reason: 'real_event_exists',
+        };
+      }
+
+      // Generate synthetic event
+      const syntheticEvent = await messageEventRepository.createForTenant(tenantId, {
+        messageId,
+        type: eventType, // no_open, no_click
+        eventAt: new Date(),
+        data: {
+          synthetic: true,
+          triggeredBy: 'timeout',
+          originalJobId: job.id?.toString(),
+          scheduledAt: job.opts?.delay
+            ? new Date(Date.now() - (job.opts.delay as number))
+            : new Date(),
+          actualFiredAt: new Date(),
+        },
+      });
+
+      logger.info('[CampaignExecutionWorker] Created synthetic event', {
+        eventId: syntheticEvent.id,
+        eventType,
+        messageId,
+        jobId: job.id,
+      });
+
+      // Trigger campaign transition processing
+      // Get the campaign to access the plan and current node
+      const campaign = await contactCampaignRepository.findByIdForTenant(campaignId, tenantId);
+
+      if (campaign && campaign.planJson && campaign.currentNodeId) {
+        const campaignPlan = campaign.planJson as any;
+
+        await campaignPlanExecutionService.processTransition({
+          tenantId,
+          campaignId,
+          eventType,
+          currentNodeId: campaign.currentNodeId,
+          plan: campaignPlan,
+          eventRef: syntheticEvent.id,
+        });
+
+        logger.info('[CampaignExecutionWorker] Campaign transition processed successfully', {
+          tenantId,
+          campaignId,
+          eventType,
+          currentNodeId: campaign.currentNodeId,
+          messageEventId: syntheticEvent.id,
+          jobId: job.id,
+        });
+      } else {
+        logger.warn(
+          '[CampaignExecutionWorker] Could not process campaign transition - missing campaign data',
+          {
+            tenantId,
+            campaignId,
+            campaignExists: !!campaign,
+            hasPlan: !!campaign?.planJson,
+            hasCurrentNode: !!campaign?.currentNodeId,
+            jobId: job.id,
+          }
+        );
+      }
+
+      return {
+        success: true,
+        syntheticEventId: syntheticEvent.id,
+      };
+    } catch (error) {
+      logger.error('[CampaignExecutionWorker] Failed to process timeout job', {
+        jobId: job.id,
+        campaignId,
+        nodeId,
+        eventType,
+        messageId,
+        tenantId,
+        error: error instanceof Error ? error.message : 'Unknown error',
+        stack: error instanceof Error ? error.stack : undefined,
+      });
+
+      return {
+        success: false,
+        reason: error instanceof Error ? error.message : 'Unknown error',
+      };
+    }
+  }
+}
